feat: modify CoffeeInformationScreen to CoffeeInformationCardScreen, refactor the respective code and create new CoffeeInformationScreen
