+++ app/src/androidTest/java/com/android/brewr/ui/e2eTest.kt
@@ -125,7 +125,7 @@ class E2ETest {
           composable(Screen.JOURNEY_RECORD) {
             JourneyRecordScreen(listJourneysViewModel, navigationActions)
           }
-          composable(EXPLORE) { ExploreScreen(sampleCoffees) }
+          composable(EXPLORE) { ExploreScreen(sampleCoffees, navigationActions) }
         }
         navigation(
             startDestination = Screen.ADD_JOURNEY,
@@ -223,5 +223,9 @@ class E2ETest {
     // check the bottomSheet and coffee shop information existence
     composeTestRule.onNodeWithTag("bottomSheet").assertIsDisplayed()
     composeTestRule.onNodeWithTag("coffeeImage").assertIsDisplayed()
+    composeTestRule.onNodeWithTag("coffeeShopName").assertIsDisplayed()
+    composeTestRule.onNodeWithTag("coffeeShopAddress").assertIsDisplayed()
+    composeTestRule.onNodeWithTag("coffeeShopHours").assertIsDisplayed()
+    composeTestRule.onNodeWithTag("coffeeShopRating").assertIsDisplayed()
   }
 }
